requirements.txt
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.11.2
gunicorn==20.1.0
psycopg2-binary==2.8.6
whitenoise==5.2.0
